+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.22 - 2021-09-15
+### Updated
+- Updated recaptcha check to return entire response
+- Updated filter options in feedback plugin aside menu
+
 ## 1.0.21 - 2021-09-09
 ### Fixed
 - Fixed issue where entryId was incorrect on the entry->createFeedback buttons